feat : adding get user registration status which return an array
