+++ controller/SelectedSubject.controller.js
@@ -65,37 +65,6 @@ async function getSelectedSubjectByNim(req, res, next) {
   }
 }
 
-async function addSelectedSubject(req, res, next) {
-  try {
-    const { userId, subjects } = req.body;
-
-    const user = await User.findById(userId);
-    if (!user) throw createError.NotFound("User not found.");
-
-    const doesUserExist = await SelectedSubject.findOne({ userId: user.id });
-    if (doesUserExist)
-      throw createError.Conflict("User already added subjects");
-
-    for (let index = 0; index < subjects.length; index++) {
-      const doesSubjectExist = await Subject.findById(subjects[index]);
-      if (!doesSubjectExist) throw createError.NotFound("Subject not found");
-    }
-
-    const selectedSubject = new SelectedSubject({ userId, subjects });
-    await selectedSubject.save();
-
-    const response = {
-      status: 200,
-      message: "success",
-      data: selectedSubject,
-    };
-
-    res.send(response);
-  } catch (error) {
-    next(error);
-  }
-}
-
 async function updateSelectedSubject(req, res, next) {
   try {
     const { userId, subjects } = req.body;
@@ -132,6 +101,5 @@ module.exports = {
   getAllSelectedSubjects,
   getSelectedSubjectById,
   getSelectedSubjectByNim,
-  addSelectedSubject,
   updateSelectedSubject,
 };
